+++ src/RcppExports.cpp
@@ -1670,6 +1670,37 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// cpp_test_nl_field_seed
+Rcpp::CharacterVector cpp_test_nl_field_seed(std::string kind, std::string tag, int n_spatial_units, double scale_factor, Rcpp::IntegerVector adj_row_ptr, Rcpp::IntegerVector adj_col_idx, int n_spatial, int nn, int cov_type, Rcpp::NumericMatrix coords, Rcpp::IntegerMatrix nn_idx, Rcpp::IntegerVector spatial_idx, int M, Rcpp::NumericMatrix phi_basis, Rcpp::NumericVector lambda_eig, std::string temporal_type, int n_times, bool cyclic, Rcpp::IntegerVector temporal_idx, int n_groups, bool with_groups);
+RcppExport SEXP _tulpa_cpp_test_nl_field_seed(SEXP kindSEXP, SEXP tagSEXP, SEXP n_spatial_unitsSEXP, SEXP scale_factorSEXP, SEXP adj_row_ptrSEXP, SEXP adj_col_idxSEXP, SEXP n_spatialSEXP, SEXP nnSEXP, SEXP cov_typeSEXP, SEXP coordsSEXP, SEXP nn_idxSEXP, SEXP spatial_idxSEXP, SEXP MSEXP, SEXP phi_basisSEXP, SEXP lambda_eigSEXP, SEXP temporal_typeSEXP, SEXP n_timesSEXP, SEXP cyclicSEXP, SEXP temporal_idxSEXP, SEXP n_groupsSEXP, SEXP with_groupsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< std::string >::type kind(kindSEXP);
+    Rcpp::traits::input_parameter< std::string >::type tag(tagSEXP);
+    Rcpp::traits::input_parameter< int >::type n_spatial_units(n_spatial_unitsSEXP);
+    Rcpp::traits::input_parameter< double >::type scale_factor(scale_factorSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type adj_row_ptr(adj_row_ptrSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type adj_col_idx(adj_col_idxSEXP);
+    Rcpp::traits::input_parameter< int >::type n_spatial(n_spatialSEXP);
+    Rcpp::traits::input_parameter< int >::type nn(nnSEXP);
+    Rcpp::traits::input_parameter< int >::type cov_type(cov_typeSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type coords(coordsSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerMatrix >::type nn_idx(nn_idxSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type spatial_idx(spatial_idxSEXP);
+    Rcpp::traits::input_parameter< int >::type M(MSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type phi_basis(phi_basisSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type lambda_eig(lambda_eigSEXP);
+    Rcpp::traits::input_parameter< std::string >::type temporal_type(temporal_typeSEXP);
+    Rcpp::traits::input_parameter< int >::type n_times(n_timesSEXP);
+    Rcpp::traits::input_parameter< bool >::type cyclic(cyclicSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type temporal_idx(temporal_idxSEXP);
+    Rcpp::traits::input_parameter< int >::type n_groups(n_groupsSEXP);
+    Rcpp::traits::input_parameter< bool >::type with_groups(with_groupsSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_test_nl_field_seed(kind, tag, n_spatial_units, scale_factor, adj_row_ptr, adj_col_idx, n_spatial, nn, cov_type, coords, nn_idx, spatial_idx, M, phi_basis, lambda_eig, temporal_type, n_times, cyclic, temporal_idx, n_groups, with_groups));
+    return rcpp_result_gen;
+END_RCPP
+}
 // cpp_test_gp_nngp_twins
 Rcpp::NumericVector cpp_test_gp_nngp_twins(Rcpp::NumericVector w, double sigma2, double phi, Rcpp::NumericMatrix coords, Rcpp::IntegerMatrix nn_idx, Rcpp::NumericMatrix nn_dist, Rcpp::NumericVector nn_neighbor_dist, Rcpp::IntegerVector nn_order, Rcpp::IntegerVector nn_order_inv, int cov_type);
 RcppExport SEXP _tulpa_cpp_test_gp_nngp_twins(SEXP wSEXP, SEXP sigma2SEXP, SEXP phiSEXP, SEXP coordsSEXP, SEXP nn_idxSEXP, SEXP nn_distSEXP, SEXP nn_neighbor_distSEXP, SEXP nn_orderSEXP, SEXP nn_order_invSEXP, SEXP cov_typeSEXP) {
@@ -4080,6 +4111,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_tulpa_cpp_test_joint_logpost_grad", (DL_FUNC) &_tulpa_cpp_test_joint_logpost_grad, 8},
     {"_tulpa_cpp_nested_laplace_multi", (DL_FUNC) &_tulpa_cpp_nested_laplace_multi, 25},
     {"_tulpa_cpp_nested_laplace_test_occupancy_likelihood", (DL_FUNC) &_tulpa_cpp_nested_laplace_test_occupancy_likelihood, 2},
+    {"_tulpa_cpp_test_nl_field_seed", (DL_FUNC) &_tulpa_cpp_test_nl_field_seed, 21},
     {"_tulpa_cpp_test_gp_nngp_twins", (DL_FUNC) &_tulpa_cpp_test_gp_nngp_twins, 10},
     {"_tulpa_cpp_test_nngp_nc_grad", (DL_FUNC) &_tulpa_cpp_test_nngp_nc_grad, 12},
     {"_tulpa_cpp_test_svc_nngp_twins", (DL_FUNC) &_tulpa_cpp_test_svc_nngp_twins, 8},